pyproject.toml
[project]
name = "sortnao"
version = "0.0.1"
description = "Illustration sorting using SauceNAO."
requires-python = ">=3.12"
dependencies = [
    "playwright==1.50.0",
    "python-dotenv==1.0.1",
    "requests==2.32.3"
]
